+++ features/ftue/impl/src/main/kotlin/io/element/android/features/ftue/impl/notifications/NotificationsOptInState.kt
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 
-package io.element.android.features.ftue.impl.notitications
+package io.element.android.features.ftue.impl.notifications
 
-sealed interface NotificationsOptInEvents {
-    data object ContinueClicked : NotificationsOptInEvents
-    data object NotNowClicked : NotificationsOptInEvents
-}
+import io.element.android.libraries.permissions.api.PermissionsState
+
+data class NotificationsOptInState(
+    val notificationsPermissionState: PermissionsState,
+    val eventSink: (NotificationsOptInEvents) -> Unit
+)
