+++ libraries/permissions/impl/src/main/kotlin/io/element/android/libraries/permissions/impl/FakePermissionStateProvider.kt
@@ -28,7 +28,7 @@ class FakePermissionStateProvider(
     private val permissionDeniedFlow = MutableStateFlow(permissionDenied)
     private val permissionAskedFlow = MutableStateFlow(permissionAsked)
 
-    fun setPermissionGranted(permission: String) {
+    fun setPermissionGranted() {
         permissionGranted = true
     }
