+++ services/toolbox/test/src/main/kotlin/io/element/android/services/toolbox/test/sdk/FakeBuildVersionSdkIntProvider.kt
@@ -16,7 +16,6 @@
 
 package io.element.android.services.toolbox.test.sdk
 
-import android.os.Build
 import io.element.android.services.toolbox.api.sdk.BuildVersionSdkIntProvider
 
 class FakeBuildVersionSdkIntProvider(
